Quizlet hacks
This quizlet hack script is different from others, as you actually INPUT the number that you WANT, not a completely unbelievable 0.5 seconds. to input the value, change the xxxx to the amount of seconds desired, followed by 3 zeros (or amount of milliseconds, depending on how you word it) also, the script does nothing with the basic value (xxxx), so please change it first.